rotation done in axis y, optimized better
